+++ src/components/MediaNotes.vue
@@ -1,16 +1,22 @@
 <template>
   <div>
     <h4>Notes</h4>
-    <markdown-editor v-model="notes" :configs="editorConfigs" ref="markdownEditor" />
+    <!-- <markdown-editor v-model="notes" :configs="editorConfigs" ref="markdownEditor" /> -->
+    <vue-editor
+      ref="editor"
+      :editorToolbar="toolbarContent"
+      v-model="notes"
+    />
   </div>
 </template>
 
 <script>
-import markdownEditor from 'vue-simplemde/src/markdown-editor'
+import { VueEditor } from 'vue2-editor'
+// import markdownEditor from 'vue-simplemde/src/markdown-editor'
 
 export default {
   components: {
-    markdownEditor
+    VueEditor
   },
   props: ['userNotes', 'mediaid', 'mediaType', 'update'],
   data () {
@@ -18,19 +24,31 @@ export default {
       notes: this.userNotes,
       id: this.mediaid,
       type: this.mediaType,
+      toolbarContent: [
+        [{ 'header': [] }],
+        ['bold', 'italic', 'underline', 'strike'],
+        ['blockquote'],
+        [{ 'list': 'ordered' }, { 'list': 'bullet' }],
+        ['clean']
+      ],
       editorConfigs: {
         spellChecker: true,
         toolbar: ['bold', 'italic', 'heading', '|', 'quote', 'unordered-list', 'ordered-list', '|', 'link', 'image']
       }
     }
   },
-  computed: {
-    simplemde () {
-      return this.$refs.markdownEditor.simplemde
-    }
-  },
+  // computed: {
+  //   simplemde () {
+  //     return this.$refs.markdownEditor.simplemde
+  //   }
+  // },
   mounted () {
-    this.simplemde.codemirror.on('blur', this.updateNotes)
+    // this.simplemde.codemirror.on('blur', this.updateNotes)
+    this.$refs.editor.quill.on('selection-change', (range, oldRange, source) => {
+      if (range === null && oldRange !== null) {
+        this.updateNotes()
+      }
+    })
   },
   watch: {
     userNotes (value) {
@@ -48,10 +66,21 @@ export default {
 <style>
 
 /* @import '~simplemde/dist/simplemde.min.css'; */
-@import '../statics/simplemde-dark-theme.min.css';
+/* @import '../statics/simplemde-dark-theme.min.css'; */
 
-.cm-spell-error {
+/* .cm-spell-error {
   background-color: rgba(255,0,0,0.2);
+} */
+.ql-picker-label {
+  color: #bbb !important;
+}
+
+.ql-fill {
+  fill: #bbb !important;
+}
+
+.ql-stroke {
+  stroke: #bbb !important;
 }
 
 </style>