+++ src/pages/index.vue
@@ -1,6 +1,5 @@
 <template>
   <q-page class="flex flex-center">
-    <img src="~assets/quasar-logo-full.svg">
     <q-btn label="Next Page" @click="$router.push('/test')" />
   </q-page>
 </template>